clean up material resolver
